% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facilityeq.R
\name{facilityeq}
\alias{facilityeq}
\title{Calculate the equilibrium of a facility transmission model}
\usage{
facilityeq(Sfun, C, Afun, R, transm, initS, mgf)
}
\arguments{
\item{Sfun}{A function that produces a matrix of state transition rates between and removal from the susceptible states for a given value of the acquisition rate}

\item{C}{A matrix of state transition rates between and removal from the colonized states}

\item{Afun}{A function that produces a matrix of transition rates from susceptible to colonized states for a given value of the acquisition rate}

\item{R}{A matrix of recovery rates: state transition rates from colonized to susceptible states}

\item{transm}{A vector of transmission rates from each colonized state}

\item{mgf}{The moment generating function characterizing a time-of-stay-dependent removal hazard}

\item{init}{A vector of admission state probabilities to each state}
}
\value{
A vector with the proportion of patients in each state at equilibrium
}
\description{
Calculate the equilibrium of a facility transmission model
}
